move the import logic to run.py
